product client added, integrated with order service

+++ product-application/src/main/kotlin/com/ecommerce/product/application/dto/ProductDto.kt
@@ -1,9 +1,11 @@
 package com.ecommerce.product.application.dto
 
+import java.math.BigDecimal
+
 data class ProductDto(
     val id: Long,
     val name: String,
-    val price: Long,
+    val price: BigDecimal,
     val quantity: Long,
     val description: String
 )
